+++ client/src/utils/index.ts
@@ -1,19 +1,13 @@
-import { clsx } from 'clsx';
-import { twMerge } from 'tailwind-merge';
-
+export * from './json';
 export * from './languages';
+export { default as cn } from './cn';
 export { default as buildTree } from './buildTree';
 export { default as getLoginError } from './getLoginError';
 export { default as cleanupPreset } from './cleanupPreset';
 export { default as validateIframe } from './validateIframe';
-export { default as getMessageError } from './getMessageError';
 export { default as buildDefaultConvo } from './buildDefaultConvo';
 export { default as getDefaultEndpoint } from './getDefaultEndpoint';
 export { default as getLocalStorageItems } from './getLocalStorageItems';
-
-export function cn(...inputs: string[]) {
-  return twMerge(clsx(inputs));
-}
 
 export const languages = [
   'java',
